reorganized directory structure to adhere to Flask documentation's example. run_me.py is not though, just for convenience
