+++ .gitignore
@@ -1,2 +1,15 @@
-__pycache__
-.venv
+.gitignore
+.venv/
+
+*.pyc
+__pycache__/
+
+instance/
+
+.pytest_cache/
+.coverage
+htmlcov/
+
+dist/
+build/
+*.egg-info/
