+++ README.md
@@ -2,7 +2,7 @@
 A Toolbox viewer and statistical analyzer for the Visual Experience Dataset developed for the CS 425/426 capstone computer science class at the University of Nevada, Reno.
 
 ## How to run
-Ensure all required packages are installed, and run run_me.py.
+Ensure all required packages are installed, and run run_me.py. this file is only for convenience; it breaks conventional Flask directory structures.
 
 ## Test User
 **Username:** test
