+++ flaskr/static/images/Nevada-Wolf-Pack-Emblem.png
