+++ flaskr/static/images/gear.png
