+++ flaskr/static/images/logout.png
