+++ flaskr/static/images/vedblogo.png
