+++ flaskr/templates/auth/register.html
@@ -6,7 +6,7 @@
 <head><!--Header, what needs to be loaded before page is made-->
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href= "{{ url_for('static',filename='styles/login-signup.css') }}">
+    <link rel="stylesheet" href= "{{ url_for('static',filename='styles/auth.css') }}">
     <title>Document</title>
     <style>
         p {
