+++ flaskr/templates/file-upload/file_upload.html
