+++ run_me.py
@@ -1,6 +1,6 @@
 # run.py
 
-from app import create_app  # Import the create_app function from app/__init__.py
+from flaskr import create_app  # Import the create_app function from app/__init__.py
 
 def run():
     # Create the Flask app using the factory function
